+++ components/api-server/test/permissions-create-only.test.js
@@ -557,19 +557,9 @@ describe('permissions create-only level', () => {
       return `${basePath}/${id}`;
     }
 
-    describe('GET /', function () {
-      it('[5FHF] should return an empty list when fetching webhooks', async function () {
-        const res = await server
-          .request()
-          .get(basePath)
-          .set('Authorization', createOnlyToken);
-        assert.equal(res.status, 200);
-        assert.equal(res.body.webhooks.length, 0);
-      });
-    });
 
     describe('CREATE /', function() {
-      it('[3AE9] should forbid creating webhooks', async function () {
+      it('[3AE9] should allow creating webhooks', async function () {
         const res = await server
           .request()
           .post(basePath)
@@ -577,7 +567,7 @@ describe('permissions create-only level', () => {
           .send({
             url: charlatan.Internet.url(),
           });
-        assert.equal(res.status, 403);
+        assert.equal(res.status, 201);
       });
     });